UPSTREAM: util/abuild: Have abuild generate the .xcompile if it doesn't exist
Previously if .xcompile was missing, abuild would silently ignore the
error. With https://review.coreboot.org/c/coreboot/+/34241 we now check
the return code so abuild started failing.

We should generate the .xcompile if it doesn't exist. The Makefile will
handle that so we include it as the first Makefile.

We then need to override the default target so we don't use the one from
the Makefile.

BUG=b:112267918
TEST=ran abuild and made sure it generated a .xcompile in the root.

Change-Id: Ic6060726598f0da7b557f6a655a23862d18f1232
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 9ff4029db9b1053b44e4fc487243017a099a49c5
Original-Change-Id: I79ded36d47b0219d0b126adff80a57be1c2bdf07
Original-Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/35137
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Martin Roth <martinroth@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1868963
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>