UPSTREAM: soc/skylake: do not rely on P2SB data to generate DRHD
The P2SB PCI device can be "hidden", which causes all sorts of
nightmares and bugs. Moreover, FSP tends to hide it, so finding
a good solution to this problem is impossible with FSP into the mix.

Since the values for IBDF and HBDF were already hardcoded as FSP
parameters, define them as macros and use these values directly to
generate the DRHD.

BUG=none
BRANCH=none
TEST=none

Change-Id: I9b93299ea7f0d185159f0222b138dc49f4ca55a7
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: ef879a8f30137a7082c365091729e31bedec69d6
Original-Change-Id: I7eb20182380b953a1842083e7a3c67919d6971b9
Original-Signed-off-by: Angel Pons <th3fanbus@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/35108
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Mimoja <coreboot@mimoja.de>
Original-Reviewed-by: Maxim Polyakov <max.senia.poliak@gmail.com>
Original-Reviewed-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1781359
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>