UPSTREAM: AUTHORS: Move src/arch/x86 copyrights into AUTHORS file
As discussed on the mailing list and voted upon, the coreboot project
is going to move the majority of copyrights out of the headers and into
an AUTHORS file.  This will happen a bit at a time, as we'll be unifying
license headers at the same time.

BUG=none
BRANCH=none
TEST=none

Change-Id: I33c011dc2747ec3ffcfe9d909c3f9463d418e260
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 20bbd81201e36ff716d51ba133d7c9bf2970e7ec
Original-Signed-off-by: Martin Roth <martin@coreboot.org>
Original-Change-Id: Ifd4329905847d9dd06de67b9a443c8ee50c0e7a7
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/35177
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1795441
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>