UPSTREAM: AUTHORS: Move src/commonlib copyrights into AUTHORS file
As discussed on the mailing list and voted upon, the coreboot project
is going to move the majority of copyrights out of the headers and into
an AUTHORS file.  This will happen a bit at a time, as we'll be unifying
license headers at the same time.

BUG=none
BRANCH=none
TEST=none

Change-Id: I01ce5750e71ad0568550d7637e959fe868bccf72
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 0443ac282757cdf46703486717b388818297181c
Original-Signed-off-by: Martin Roth <martin@coreboot.org>
Original-Change-Id: I4c9351652d81040cc4e7b85bdd1ba85709a74192
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/35178
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1795782
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>