UPSTREAM: AUTHORS: Move src/console copyrights into AUTHORS file
As discussed on the mailing list and voted upon, the coreboot project
is going to move the majority of copyrights out of the headers and into
an AUTHORS file.  This will happen a bit at a time, as we'll be unifying
license headers at the same time.

BUG=none
BRANCH=none
TEST=none

Change-Id: I5a674d94191a87f2ecc2935b384c6de55773e687
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 65244a7f66af89c4fb4e8f00466b289d71da4771
Original-Signed-off-by: Martin Roth <martin@coreboot.org>
Original-Change-Id: I2f350cc3008b17516b5a42cdf07e28d2da5995e9
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/35179
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1795783
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>