UPSTREAM: AUTHORS: Move src/cpu/amd copyrights into AUTHORS file
As discussed on the mailing list and voted upon, the coreboot project
is going to move the majority of copyrights out of the headers and into
an AUTHORS file.  This will happen a bit at a time, as we'll be unifying
license headers at the same time.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ie81e3ead911d593c34498d1cd51e24154f74b258
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 97ea346fdf573faa3cf65a18edec18c32c73d306
Original-Signed-off-by: Martin Roth <martin@coreboot.org>
Original-Change-Id: I77275adb7c15b242e319805b8a60b7755fa25db5
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/35180
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1795784
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>