UPSTREAM: soc/intel/common/block/cse: Move me_read_config32() to common code
me_read_config32() is defined in multiple places, move it to common
location. Also, this function is usually used for reading HFSTS
registers, hence move the HFSTS register definitions to common location.

Also add a funtion to check if the CSE device has been enabled in the
devicetree and it is visible on the bus. This API can be used by
the caller to check before initiating any HECI communication.

TEST=Verified reading HFSTS registers on CML RVP & Hatch board

Change-Id: I11114810e23ec4a412fd3e6b0d0e524b32134c71
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 2cc66916e5d5d5b0d2e92e180bae7ac64d30cbac
Original-Change-Id: Icdbfb6b30a007d469b5e018a313c14586addb130
Original-Signed-off-by: Sridhar Siricilla <sridhar.siricilla@intel.com>
Original-Signed-off-by: Rizwan Qureshi <rizwan.qureshi@intel.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/35225
Original-Reviewed-by: Subrata Banik <subrata.banik@intel.com>
Original-Reviewed-by: Aamir Bohra <aamir.bohra@intel.com>
Original-Reviewed-by: V Sowmya <v.sowmya@intel.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1801370
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>