UPSTREAM: AUTHORS: Move src/cpu copyrights into AUTHORS file
As discussed on the mailing list and voted upon, the coreboot project
is going to move the majority of copyrights out of the headers and into
an AUTHORS file.  This will happen a bit at a time, as we'll be unifying
license headers at the same time.

BUG=none
BRANCH=none
TEST=none

Change-Id: I8cc7abc4036537f02c743eac651e63fc85d53e59
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: d57ace259ad89ccba0b2e72036d84ff4016b1d32
Original-Signed-off-by: Martin Roth <martin@coreboot.org>
Original-Change-Id: Id6070fb586896653a1e44951a6af8f42f93b5a7b
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/35184
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1795786
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>