UPSTREAM: AUTHORS: Move src/device copyrights into AUTHORS file
As discussed on the mailing list and voted upon, the coreboot project
is going to move the majority of copyrights out of the headers and into
an AUTHORS file.  This will happen a bit at a time, as we'll be unifying
license headers at the same time.

Additional cleanup - Unify "Inc" to "Inc." and "LLC." to "LLC"

BUG=none
BRANCH=none
TEST=none

Change-Id: I19d08d1b0abc4c0e48ea29972b5b2883d320d3a7
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 99f83bbad4a7e060e05c3bb26203f92c1817445a
Original-Signed-off-by: Martin Roth <martin@coreboot.org>
Original-Change-Id: Ie03a3ce1f6085494bd5f38da76e2467970cf301a
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/35430
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1809141
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Sean Abraham <seanabraham@chromium.org>