UPSTREAM: mb/google/hatch: Remove GPIO_DRIVER from pen eject GPIO configuration
A closer read of the EDS indicates that when GPIO Driver mode is
selected, GPIO input event updates are limited to GPI_STS only.
GPI_GPE_STS updates are therefore masked, and we don't want to enable
this behavior.  It masks the GPE and does not allow us to see this GPE
as a wake source, obscuring the reason that the system woke up.

Also switch the IRQ from level-triggered to edge-triggered,
otherwise the system will auto-wake from any sleep state when the
pen is ejected from the garage.

BUG=b:132981083
BRANCH=none
TEST=Wake up system from S0ix using pen eject, verify that mosys
eventlog shows GPE#8 as the S0ix wakeup source.  Wake up system from S3
via pen eject, and verify that the wakeup source shows as GPE#8.

Change-Id: I97b26f148bf6c78429d7d02f87d3876a273794b9
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 954111695cb7e27647ed9c150cac2f77d39c1e09
Original-Change-Id: If017e12e23134f5cfed7cbb6047cc9badd9bf7e8
Original-Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/35459
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1816399
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>