UPSTREAM: soc/qualcomm/ipq40xx: Remove unnecessary allocation
The bus variable doesn't live outside the scope of this function, and is
only used as a convenient way for passing the pointers to all the
sub-functions, so it doesn't need to be allocated. Put it on the stack
instead. A similar fix for ipq806x was done in 0f33d8c29a
(soc/qualcomm/ipq806x: Remove unnecessary allocation).

BUG=none
BRANCH=none
TEST=none

Change-Id: I6943efd9584057404cf686f42855aa9ce325bb3c
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 145eb479a4e9f2c32f10d8afdfea3393395dea9b
Original-Change-Id: Ibb1129b92e38a105e100f59e03d107de340b925c
Original-Signed-off-by: Jacob Garber <jgarber1@ualberta.ca>
Original-Found-by: Coverity CID 1294801
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/35464
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1816402
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>