UPSTREAM: AUTHORS: Move src/drivers/[a*-i*] copyrights into AUTHORS file
As discussed on the mailing list and voted upon, the coreboot project
is going to move the majority of copyrights out of the headers and into
an AUTHORS file.  This will happen a bit at a time, as we'll be unifying
license headers at the same time.

Updated Authors file is in a separate commit.

BUG=none
BRANCH=none
TEST=none

Change-Id: I28e2a261c4e807f8937d1de478c475997baab682
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: cddd6008f9434c013efcd3a0533265e54b245d75
Original-Signed-off-by: Martin Roth <martin@coreboot.org>
Original-Change-Id: I1acea8c975d14904b7e486dc57a1a67480a6ee6e
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/36178
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1874530
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>