UPSTREAM: cpu/amd/pi/00730F01: Clean up microcode update from __PRE_RAM__
It is only called in ramstage. Even if it was called in
romstage, execution flow is such that BSP and AP CPUs
should not be able to enter update_microcode() routine
concurrently.

Also the Kconfig guarding the spin_lock() calls are not
selected nor are the lock variables declared for these
platforms.

BUG=none
BRANCH=none
TEST=none

Change-Id: I6747b20c4e4b22e1c944954525904a46cfe87eff
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 3de8f8c137f336ec2b96114303cee873ac636762
Original-Change-Id: I1c2e106f10e8420e942b3ed082c677c0db95557c
Original-Signed-off-by: Kyösti Mälkki <kyosti.malkki@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/35586
Original-Reviewed-by: Michał Żygowski <michal.zygowski@3mdeb.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1827172
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>