UPSTREAM: coreinfo: Add support for link time optimization
This introduces a Kconfig option for compiling coreinfo with LTO.
This option can be used independently of LTO in libpayload, though will
benefit most if that is enabled as well. If both are enabled, the
final size of coreinfo.elf is reduced from 95 KiB to 92 KiB.

Tested in QEMU and on Thinkpad T500.

BUG=none
BRANCH=none
TEST=none

Change-Id: I664a229f8b068594123535564f8e88780712e4ec
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 5e6c30bc421757a11ec6f7c1f730bf36c7de1b57
Original-Change-Id: I6feacdb911b52b946869bff369e03dcf72897c9f
Original-Signed-off-by: Jacob Garber <jgarber1@ualberta.ca>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/38293
Original-Reviewed-by: Michael Niewöhner <foss@mniewoehner.de>
Original-Reviewed-by: Martin Roth <martinroth@google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2516462
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>