UPSTREAM: AUTHORS: Move src/drivers/[l*-v*] copyrights into AUTHORS file
As discussed on the mailing list and voted upon, the coreboot project
is going to move the majority of copyrights out of the headers and into
an AUTHORS file.  This will happen a bit at a time, as we'll be unifying
license headers at the same time.

Updated Authors file is in a separate commit.

BUG=none
BRANCH=none
TEST=none

Change-Id: Iee5bddde40bdd8f57854579e27b7abaa99e5df06
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: effaf8fcb8af5f2a0a3355977de22f402c780da0
Original-Signed-off-by: Martin Roth <martin@coreboot.org>
Original-Change-Id: Ia0a07df6ca1fdaa2837ce8839057057cbd44d157
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/36181
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1874531
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>