UPSTREAM: Updated AUTHORS file for src/drivers
As discussed on the mailing list and voted upon, the coreboot project
is going to move the majority of copyrights out of the headers and into
an AUTHORS file.  This will happen a bit at a time, as we'll be unifying
license headers at the same time.

BUG=none
BRANCH=none
TEST=none

Change-Id: I83cec47f7bf9a37c4ab0a1aa8359df09ea6f567f
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: f88bad8331a8575bfdc9080dfab5a93f565de3b5
Original-Signed-off-by: Martin Roth <martin@coreboot.org>
Original-Change-Id: I446cd267dec2b6c0f7742a6834c7ee950528ad5f
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/36182
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1874532
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>